+++ src/main/java/ru/job4j/cars/service/api/Service.java
@@ -6,11 +6,11 @@ public interface Service<T> {
 
     T add(T entity);
 
-    boolean replace(int id, T entity);
+    boolean replace(long id, T entity);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     List<T> findAll();
 
-    T findById(int id);
+    T findById(long id);
 }
